pyproject.toml
[tool.poetry]
name = "fastscaledb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
fastapi-mqtt = "^2.2.0"
uvicorn = {version = "^0.30.6", extras = ["standard"]}
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.1"
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
